+++ public/usage-examples/geometry/point_in_circle/point_in_circle-1-simple.txt
@@ -1,0 +1,3 @@
+#### Point in tcircle
+
+The following code shows examples of using [Point In Circle](/api/geometry/#point-in-circle) to show how check if the point in circle or not.
