--- public/usage-examples/geometry/point_in_triangle/point_in_triangle-1-simple.txt
@@ -1,3 +1,0 @@
-#### Point in triangle
-
-The following code shows examples of using [Point In Triangle](/api/geometry/#point-in-triangle) to show how check if the point in triangle or not.
